Reordered folder structure
