+++ practice-app/api/views.py
@@ -1,6 +1,9 @@
-from rest_framework.views import APIView
-from rest_framework.response import Response
 
+from rest_framework.response import Response
+from rest_framework.views import APIView
+from rest_framework import status as st
+
+from . import serializers as ls
 
 class HelloWorldView(APIView):
     def get(self, request):
@@ -14,3 +17,10 @@ class HelloWorldView(APIView):
         return Response({
             'message': f'Hello, {name}.'
         })
+
+class LoginAPIView(APIView):
+    def post(self, *args, **kwargs):
+        data = self.request.data
+        serializer = ls.LoginSerializer(data=data)
+        serializer.is_valid(raise_exception=True)
+        return Response(serializer.data, status=st.HTTP_200_OK)
