--- practice-app/login/urls.py
@@ -1,6 +1,0 @@
-from django.conf.urls import url
-from login.views import *
-
-urlpatterns = [
-    url(r'^login$', LoginAPIView.as_view()),
-]
